+++ docs/troubleshooting/meta.ts
@@ -3,6 +3,6 @@ import { defineMeta } from "blume";
 export default defineMeta({
   title: "Troubleshooting",
   icon: "life-buoy",
-  order: 6,
+  order: 7,
   pages: ["index", "blocked-plugins", "netcup-mariadb"],
 });